math done, need to add tests

+++ include/atmosphere.hpp
@@ -4,12 +4,12 @@
 #include <cstdint>
 
 struct atmosphericData {
-    uint8_t temp; // decimals better, but limited memory
-    uint8_t elevation;
-    uint8_t vWind;
-    uint8_t phiWind; // angle of wind w.r.t Y-axis [-180, 180]; ex. 0deg -> CF, 45deg -> RF
-    uint8_t hWind; // height where wind acts
-    uint8_t relHumidity;
+    float temp; // decimals better, but limited memory
+    float elevation;
+    float vWind;
+    float phiWind; // angle of wind w.r.t Y-axis [-180, 180]; ex. 0deg -> CF, 45deg -> RF
+    float hWind; // height where wind acts
+    float relHumidity;
     float pressure; // Hg
 };
